% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arctic}
\alias{arctic}
\title{Arctic Lake}
\format{
A \code{\link{data.frame}} with 4 variables:
\describe{
\item{sand}{Sand content (percent).}
\item{silt}{Silt content (percent).}
\item{clay}{Clay content (percent).}
\item{depth}{Water depth (m).}
}
}
\usage{
arctic
}
\description{
Sand, silt, clay compositions of 39 sediment samples at different water
depths in an Arctic lake.
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall.
}
\seealso{
Other datasets: 
\code{\link{hongite}},
\code{\link{slides}}
}
\concept{datasets}
\keyword{datasets}
