% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{augment}
\alias{augment}
\alias{augment,CompositionMatrix-method}
\alias{augment,LogRatio-method}
\title{Augment Data with Extra Columns}
\usage{
\S4method{augment}{CompositionMatrix}(x, ...)

\S4method{augment}{LogRatio}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{CompositionMatrix}} or \code{\linkS4class{LogRatio}} object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Adds columns from the original data.
}
\examples{
## Data from Aitchison 1986
data("arctic")

## Coerce to compositional data
coda <- as_composition(arctic, parts = 1:3)

## Isometric log-ratio
ilr <- transform_ilr(coda)

## Add extra column
augment(ilr)
}
\seealso{
Other compositional data tools: 
\code{\link{as_amounts}()},
\code{\link{as_composition}()}
}
\author{
N. Frerebeau
}
\concept{compositional data tools}
