% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/transform.R
\docType{methods}
\name{transform_clr}
\alias{transform_clr}
\alias{transform_clr-method}
\alias{transform_clr,CompositionMatrix-method}
\title{Centered Log-Ratios (CLR)}
\usage{
transform_clr(object, ...)

\S4method{transform_clr}{CompositionMatrix}(object, weights = FALSE)
}
\arguments{
\item{object}{A \linkS4class{CompositionMatrix} object.}

\item{...}{Currently not used.}

\item{weights}{A \code{\link{logical}} scalar: sould a varying weight be used. If
\code{FALSE} (the default), equally-weighted parts are used. Alternatively, a
positive \code{\link{numeric}} vector of weights can be specified.}
}
\value{
A \linkS4class{CLR} object.
}
\description{
Computes CLR transformation.
}
\examples{
## Coerce to chemical data
data("hongite")
coda <- as_composition(hongite)

## Centered log-ratio
clr <- transform_clr(coda)

## Additive log-ratio
alr <- transform_alr(coda)

## Isometric log-ratio
ilr1 <- transform_ilr(coda)
ilr2 <- transform_pivot(coda)
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall. \doi{10.1007/978-94-009-4109-0}.

Greenacre, M. J. (2019). \emph{Compositional Data Analysis in Practice}.
Boca Raton: CRC Press.
}
\seealso{
Other log-ratio transformations: 
\code{\link{LogRatio}},
\code{\link{transform_alr}()},
\code{\link{transform_ilr}()},
\code{\link{transform_inverse}()},
\code{\link{transform_pivot}()}
}
\author{
N. Frerebeau
}
\concept{log-ratio transformations}
