% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{closure}
\alias{closure}
\alias{closure-method}
\alias{closure,numeric-method}
\alias{closure,matrix-method}
\title{Closure Operation}
\usage{
closure(object, ...)

\S4method{closure}{numeric}(object, total = 1)

\S4method{closure}{matrix}(object, total = 1)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector or matrix.}

\item{...}{Currently not used.}

\item{total}{A \link{numeric} vector specifying the total amount to which the
compositions should be closed (defaults to 1).}
}
\value{
A \code{\link{numeric}} vector or matrix (same as \code{object}).
}
\description{
Closes compositions to sum up to 1.
}
\section{Missing Values Policy}{

Missing values will be omitted from the calculations.
}

\examples{
## Create a count matrix
A1 <- matrix(data = as.numeric(sample(1:100, 100, TRUE)), nrow = 20)

## Coerce to compositions
B <- as_composition(A1)

## Row sums are internally stored before coercing to relative frequencies
## (use get_totals() to retrieve these values)
## This allows to restore the source data
A2 <- as_count(B)

## Coerce to an S3 data.frame
X <- data.frame(B)
head(X)
}
\seealso{
Other classes: 
\code{\link{CompositionMatrix-class}},
\code{\link{LogRatio-class}},
\code{\link{LogicalMatrix-class}},
\code{\link{NumericMatrix-class}},
\code{\link{OutlierIndex-class}},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{classes}
