% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{as_count}
\alias{as_count-method}
\alias{as_composition}
\alias{as_composition-method}
\alias{as_composition,matrix-method}
\alias{as_composition,data.frame-method}
\alias{as_count,CompositionMatrix-method}
\alias{as_long,CompositionMatrix-method}
\title{Coerce}
\usage{
as_count(from, ...)

as_composition(from, ...)

\S4method{as_composition}{matrix}(from)

\S4method{as_composition}{data.frame}(from, samples = NULL, groups = NULL)

\S4method{as_count}{CompositionMatrix}(from)

\S4method{as_long}{CompositionMatrix}(from, factor = FALSE, reverse = FALSE)
}
\arguments{
\item{from}{An object to be coerced.}

\item{...}{Currently not used.}

\item{samples}{XXX.}

\item{groups}{XXX.}

\item{factor}{A \code{\link{logical}} scalar: should character string be
coerced to \code{\link{factor}}? Default to \code{FALSE}, if \code{TRUE} the original ordering is
preserved.}

\item{reverse}{A \code{\link{logical}} scalar: should the order of factor levels be
reversed? Only used if \code{factor} is \code{TRUE}. Useful for plotting.}
}
\value{
A coerced object.
}
\description{
Coerces an object to a \code{CompositionMatrix} object.
}
\details{
The following methods coerce an object to a \verb{*Matrix} object:

\tabular{lll}{
\strong{Method} \tab \strong{Target} \tab \strong{Details} \cr
\code{as_count()} \tab \code{\link{matrix}} \tab absolute frequency data \cr
\code{as_composition()} \tab \linkS4class{CompositionMatrix} \tab relative frequency data \cr
}

\tabular{lll}{
\strong{Method} \tab \strong{Target} \tab \strong{Details} \cr
\code{as_long()} \tab \code{\link{data.frame}} \tab long S3 data frame \cr
\code{as_features()} \tab \code{\link{data.frame}} \tab wide S3 data frame \cr
}

\code{as_features()} converts a \verb{*Matrix} object to a collection of features:
a \code{\link{data.frame}} with all informations as extra columns (result may differ
according to the class of \code{from}).

The \code{CompositionMatrix} class has special slots:
\itemize{
\item \code{samples} for replicated measurements/observation,
\item \code{groups} to group data by site/area.
}

When coercing a \code{data.frame} to a \linkS4class{CompositionMatrix} object, an
attempt is made to automatically assign values to these slots by mapping
column names (case insensitive, plural insensitive). This behavior can be
disabled by setting \code{options(nexus.autodetect = FALSE)} or overrided by
explicitly specifying the columns to be used in \code{as_composition()}.
}
\examples{
## Create a count matrix
A1 <- matrix(data = as.numeric(sample(1:100, 100, TRUE)), nrow = 20)

## Coerce to compositions
B <- as_composition(A1)

## Row sums are internally stored before coercing to relative frequencies
## (use get_totals() to retrieve these values)
## This allows to restore the source data
A2 <- as_count(B)

## Coerce to an S3 data.frame
X <- data.frame(B)
head(X)
}
\seealso{
Other classes: 
\code{\link{CompositionMatrix-class}},
\code{\link{NumericMatrix-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
