% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/initialize.R
\docType{class}
\name{CompositionMatrix-class}
\alias{CompositionMatrix-class}
\alias{.CompositionMatrix}
\alias{CompositionMatrix}
\title{Compositional Matrix}
\usage{
CompositionMatrix(data = 0, nrow = 1, ncol = 1, byrow = FALSE, dimnames = NULL)
}
\arguments{
\item{data}{an optional data vector (including a list or
    \code{\link[base]{expression}} vector).  Non-atomic classed \R objects are
    coerced by \code{\link[base]{as.vector}} and all attributes discarded.}

\item{nrow}{the desired number of rows.}

\item{ncol}{the desired number of columns.}

\item{byrow}{logical. If \code{FALSE} (the default) the matrix is
    filled by columns, otherwise the matrix is filled by rows.}

\item{dimnames}{A \code{\link[base]{dimnames}} attribute for the matrix:
    \code{NULL} or a \code{list} of length 2 giving the row and column
    names respectively.  An empty list is treated as \code{NULL}, and a
    list of length one as row names.  The list can be named, and the
    list names will be used as names for the dimensions.}
}
\description{
An S4 class to represent compositional data.
}
\section{Slots}{

\describe{
\item{\code{totals}}{A \code{\link{numeric}} vector to store the absolute row sums (before
the closure of the compositions).}

\item{\code{samples}}{A \code{\link{character}} vector to store the sample identifiers
(allows duplicates in case of replicated measurements).}

\item{\code{groups}}{A \code{\link{character}} vector to store the group names (if any).}
}}

\note{
This class inherits from \code{\link{matrix}}.
}
\examples{
## Create a compositional matrix
A <- CompositionMatrix(data = sample(1:100, 100, TRUE), nrow = 20)

## Access
dim(A) # Get the matrix dimensions
row(A) # Get the row indexes
col(A, as.factor = TRUE) # Get the column indexes
nrow(A) # Get the number of rows
ncol(A) # Get the number of columns
dimnames(A) # Get the dimension names
rownames(A) <- LETTERS[1:20] # Set the row names
rownames(A) # Get the rownames
colnames(A) <- letters[21:25] # Set the column names
colnames(A) # Get the column names

## Subset
A[[1, 1]] # Get the first value
A[1] # Get the first value
A[, ] # Get all values
A[1, , drop = FALSE] # Get the first row
A[, 1:3] # Get the first three column
}
\seealso{
\code{\link[=as_composition]{as_composition()}}

Other classes: 
\code{\link{NumericMatrix-class}},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{classes}
