% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{covariance}
\alias{covariance}
\alias{var,CompositionMatrix-method}
\alias{cov,CompositionMatrix,ANY-method}
\alias{cov,CompositionMatrix-method}
\title{Variance and Covariance}
\usage{
\S4method{var}{CompositionMatrix}(x)

\S4method{cov}{CompositionMatrix,ANY}(x)
}
\arguments{
\item{x}{A \linkS4class{CompositionMatrix} object.}
}
\value{
A \code{\link{matrix}}.
}
\description{
\itemize{
\item \code{var()} computes the log-ratio variance matrix.
\item \code{cov()} computes the log-ratio covariance matrix.
}
}
\examples{
## Coerce to compositional data
data("hongite")
coda <- as_composition(hongite)

## Variance matrix
var(coda)

## Covariance matrix
cov(coda)

## Variation matrix
variation(coda)
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall, p. 64-91. \doi{10.1007/978-94-009-4109-0}.

Greenacre, M. J. (2019). \emph{Compositional Data Analysis in Practice}.
Boca Raton: CRC Press.
}
\seealso{
Other statistics: 
\code{\link{aggregate}()},
\code{\link{dist}()},
\code{\link{mahalanobis}()},
\code{\link{mean}()},
\code{\link{variation}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
