% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coxite}
\alias{coxite}
\title{Coxite Mineralogy}
\format{
A \code{\link{data.frame}} with 5 variables (minerals):
\describe{
\item{\code{A}}{Albite (weight percent).}
\item{\code{B}}{Blandite (weight percent).}
\item{\code{C}}{Cornite (weight percent).}
\item{\code{D}}{Daubite (weight percent).}
\item{\code{E}}{Endite (weight percent).}
\item{\code{depth}}{Depth (meter).}
\item{\code{porosity}}{Porosity (percent).}
}
}
\usage{
coxite
}
\description{
Mineral compositions, depths and porosity of 25 specimens of coxite.
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall.
}
\seealso{
Other datasets: 
\code{\link{arctic}},
\code{\link{boxite}},
\code{\link{hongite}},
\code{\link{kongite}},
\code{\link{lava}},
\code{\link{predator}},
\code{\link{slides}}
}
\concept{datasets}
\keyword{datasets}
