% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/describe.R
\docType{methods}
\name{describe}
\alias{describe}
\alias{describe,CompositionMatrix-method}
\alias{describe,GroupedComposition-method}
\title{Data Description}
\usage{
\S4method{describe}{CompositionMatrix}(x)

\S4method{describe}{GroupedComposition}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{CompositionMatrix}} object.}
}
\value{
\code{describe()} is called for its side-effects. Invisibly returns \code{x}.
}
\description{
Describes an object.
}
\examples{
## Data from Aitchison 1986
data("slides")

## Coerce to compositional data
coda <- as_composition(slides, groups = 2)

## Quick description
describe(coda)
}
\author{
N. Frerebeau
}
\concept{data summaries}
