% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{distance}
\alias{distance}
\alias{dist,CompositionMatrix-method}
\title{Distances}
\usage{
\S4method{dist}{CompositionMatrix}(x, method = "aitchison", diag = FALSE, upper = FALSE)
}
\arguments{
\item{x}{A \linkS4class{CompositionMatrix} object.}

\item{method}{A \code{\link{character}} string specifying the distance measure to be
used. This must be one of "\code{aitchison}".
Any unambiguous substring can be given.}

\item{diag}{A \code{\link{logical}} scalar indicating whether the diagonal of the
distance matrix should be printed.}

\item{upper}{A \code{\link{logical}} scalar indicating whether the upper triangle of
the distance matrix should be printed.}
}
\value{
A \code{\link{dist}} object.
}
\description{
Computes the log-ratio variance matrix.
}
\examples{
## Coerce to compositonal data
data("hongite")
coda <- as_composition(hongite)

## Variance matrix
var(coda)

## Covariance matrix
cov(coda)

## Variation matrix
variation(coda)

## Distance
dist(coda)
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall, p. 64-91. \doi{10.1007/978-94-009-4109-0}.

Greenacre, M. J. (2019). \emph{Compositional Data Analysis in Practice}.
Boca Raton: CRC Press.
}
\seealso{
Other statistics: 
\code{\link{covariance}},
\code{\link{variation}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
