% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/chemistry.R
\docType{methods}
\name{element}
\alias{element}
\alias{element_major}
\alias{element_major-method}
\alias{element_minor}
\alias{element_minor-method}
\alias{element_trace}
\alias{element_trace-method}
\alias{element_major,CompositionMatrix-method}
\alias{element_minor,CompositionMatrix-method}
\alias{element_trace,CompositionMatrix-method}
\title{Chemical Elements and Compounds}
\usage{
element_major(object, ...)

element_minor(object, ...)

element_trace(object, ...)

\S4method{element_major}{CompositionMatrix}(object, min = 1/100, max = Inf)

\S4method{element_minor}{CompositionMatrix}(object, min = 0.1/100, max = 1/100)

\S4method{element_trace}{CompositionMatrix}(object, min = -Inf, max = 0.1/100)
}
\arguments{
\item{object}{A \code{\linkS4class{CompositionMatrix}} object.}

\item{...}{Currently not used.}

\item{min}{A length-one \code{\link{numeric}} vector specifying the lower bound for
element identification.}

\item{max}{A length-one \code{\link{numeric}} vector specifying the upper bound for
element identification.}
}
\value{
A \code{\link{logical}} vector.
}
\description{
Identify major, minor and traces elements in a compositional data matrix.
}
\details{
There is no definite classification of what are the major, minor and trace
elements are. By default, the following rule of thumb is used:
\describe{
\item{major elements}{The major elements are those that define the material
under study. Major elements usually have concentrations of above 1\%.}
\item{minor elements}{Minor elements usually have concentrations between
1\% and 0.1\%}
\item{trace elements}{Trace elements usually have concentrations of less
than 0.1\%.}
}
}
\examples{
## Data from Aitchison 1986
data("hongite")

## Coerce to compositional data
coda <- as_composition(hongite)

## Bar plot
barplot(coda)

## Data from Day et al. 2011
data("kommos", package = "folio")
kommos <- remove_NA(kommos, margin = 1) # Remove cases with missing values
coda <- as_composition(kommos, groups = 1) # Coerce to compositional data

## Use ceramic types for grouping
barplot(coda, order_columns = TRUE)

## Display only minor elements
barplot(coda, subset = element_minor(coda), order_columns = TRUE)
}
\author{
N. Frerebeau
}
\concept{tools}
