% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{gmean}
\alias{gmean}
\title{Geometric Mean}
\usage{
gmean(x, trim = 0, na.rm = FALSE, zero.rm = TRUE)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector.}

\item{trim}{A length-one \code{\link{numeric}} vector specifying the fraction (0 to 0.5)
of observations to be trimmed from each end of \code{x} before the mean is
computed.}

\item{na.rm}{A \code{\link{logical}} scalar: should \code{NA} values be stripped before the
computation proceeds?}

\item{zero.rm}{A \code{\link{logical}} scalar: should zeros be stripped before the
computation proceeds?}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Geometric Mean
}
\keyword{internal}
