% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/graph.R
\docType{methods}
\name{graph}
\alias{graph}
\alias{graph-method}
\alias{graph,LR-method}
\alias{graph,ALR-method}
\alias{graph,ILR-method}
\title{Graph of Log-ratios}
\usage{
graph(object, ...)

\S4method{graph}{LR}(object)

\S4method{graph}{ALR}(object)

\S4method{graph}{ILR}(object)
}
\arguments{
\item{object}{A \linkS4class{LogRatio} object.}

\item{...}{Currently not used.}
}
\value{
An \pkg{igraph} graph object.
}
\description{
Produces a graph of log-ratios.
}
\examples{
## Coerce to compositonal data
data("hongite")
coda <- as_composition(hongite)

## Pairwise log-ratio
lr <- transform_lr(coda)
lr_graph <- graph(lr)
plot(lr_graph)

## Additive log-ratio
alr <- transform_alr(coda)
alr_graph <- graph(alr)
plot(alr_graph)

## Isometric log-ratio
ilr <- transform_ilr(coda)
ilr_graph <- graph(ilr)
plot(ilr_graph)

plr <- transform_plr(coda)
plr_graph <- graph(plr)
plot(plr_graph)
}
\seealso{
Other plot methods: 
\code{\link{plot_coda}},
\code{\link{plot_logratio}}
}
\author{
N. Frerebeau
}
\concept{plot methods}
