% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/group.R
\docType{methods}
\name{group}
\alias{group}
\alias{group-method}
\alias{group<-}
\alias{any_assigned}
\alias{any_assigned-method}
\alias{is_assigned}
\alias{is_assigned-method}
\alias{is_assigned,CompositionMatrix-method}
\alias{is_assigned,LogRatio-method}
\alias{any_assigned,CompositionMatrix-method}
\alias{any_assigned,LogRatio-method}
\alias{group,CompositionMatrix-method}
\alias{group,LogRatio-method}
\alias{group,OutlierIndex-method}
\alias{group<-,CompositionMatrix-method}
\title{Working With Groups}
\usage{
group(object)

group(object) <- value

any_assigned(object)

is_assigned(object)

\S4method{is_assigned}{CompositionMatrix}(object)

\S4method{is_assigned}{LogRatio}(object)

\S4method{any_assigned}{CompositionMatrix}(object)

\S4method{any_assigned}{LogRatio}(object)

\S4method{group}{CompositionMatrix}(object)

\S4method{group}{LogRatio}(object)

\S4method{group}{OutlierIndex}(object)

\S4method{group}{CompositionMatrix}(object) <- value
}
\arguments{
\item{object}{An object from which to get or set \code{groups}.}

\item{value}{A possible value for the \code{groups} of \code{x}.}
}
\value{
\itemize{
\item \code{group() <- value} returns an object of the same sort as \code{x} with the new
group names assigned.
\item \code{group()} returns a \code{\link{character}} vector giving the group names of \code{x}.
\item \code{any_assigned()} returns a \code{\link{logical}} scalar specifying whether or not \code{x}
has groups.
\item \code{is_assigned()} returns a \code{\link{logical}} vector specifying whether or not an
observation belongs to a group.
}
}
\description{
Retrieves or defines the groups to which the observations belong.
}
\examples{
## Data from Aitchison 1986
data("slides")
head(slides)

## Coerce to compositional data
coda <- as_composition(slides, groups = 2)

group(coda)
}
\seealso{
Other mutators: 
\code{\link{mutators}},
\code{\link{totals}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
