% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{identifiers}
\alias{identifiers}
\alias{get_identifiers}
\alias{get_identifiers-method}
\alias{set_identifiers<-}
\alias{set_identifiers-method}
\alias{get_identifiers,CompositionMatrix-method}
\alias{get_identifiers,LogRatio-method}
\alias{get_identifiers,OutlierIndex-method}
\alias{set_identifiers<-,CompositionMatrix-method}
\alias{set_identifiers,CompositionMatrix-method}
\title{Unique Identifiers}
\usage{
get_identifiers(x)

set_identifiers(x) <- value

\S4method{get_identifiers}{CompositionMatrix}(x)

\S4method{get_identifiers}{LogRatio}(x)

\S4method{get_identifiers}{OutlierIndex}(x)

\S4method{set_identifiers}{CompositionMatrix}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set \code{codes}.}

\item{value}{A possible value for the \code{codes} of \code{x}.}
}
\value{
\itemize{
\item \code{set_identifiers()} returns an object of the same sort as \code{x} with the new
identifiers assigned.
\item \code{get_identifiers()} returns a \code{\link{character}} vector giving the unique
identifiers of \code{x}.
}
}
\description{
Retrieves or defines the unique identifier (eg. laboratory codes) of each
observation.
}
\details{
See \code{vignette("nexus")}.
}
\seealso{
Other mutators: 
\code{\link{groups}},
\code{\link{samples}},
\code{\link{split}()},
\code{\link{subset}()},
\code{\link{totals}}
}
\author{
N. Frerebeau
}
\concept{mutators}
