% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{is_grouped}
\alias{is_grouped}
\title{Check if an Object is Grouped}
\usage{
is_grouped(object)
}
\arguments{
\item{object}{An \R object.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
Check if an Object is Grouped
}
\seealso{
Other grouping methods: 
\code{\link{group}()},
\code{\link{group_extract}()},
\code{\link{group_metadata}},
\code{\link{group_split}()}
}
\author{
N. Frerebeau
}
\concept{grouping methods}
