% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{labels}
\alias{labels}
\alias{labels,CompositionMatrix-method}
\alias{labels,LogRatio-method}
\title{Labels}
\usage{
\S4method{labels}{CompositionMatrix}(object, ...)

\S4method{labels}{LogRatio}(object, ...)
}
\arguments{
\item{object}{An object from which to get the labels.}

\item{...}{Currently not used.}
}
\description{
Find a suitable set of labels from an object.
}
\seealso{
Other mutators: 
\code{\link{as.data.frame}()},
\code{\link{totals}()},
\code{\link{weights}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
