% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lava}
\alias{lava}
\title{Skye Lavas Compositions}
\format{
A \code{\link{data.frame}} with 3 variables (percent):
\describe{
\item{\code{A}}{Na2O + K2O.}
\item{\code{F}}{Fe2O3.}
\item{\code{M}}{MgO.}
}
}
\source{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall. \doi{10.1007/978-94-009-4109-0}.
}
\usage{
lava
}
\description{
Chemical compositions of 23 aphyric Skye lavas.
}
\seealso{
Other datasets: 
\code{\link{arctic}},
\code{\link{boxite}},
\code{\link{coxite}},
\code{\link{hongite}},
\code{\link{kongite}},
\code{\link{predator}},
\code{\link{slides}}
}
\concept{datasets}
\keyword{datasets}
