% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LogicalMatrix-class}
\alias{LogicalMatrix-class}
\alias{.LogicalMatrix}
\title{Logical Matrix}
\description{
S4 classes that represent a \eqn{m \times p}{m x p} logical matrix.
}
\note{
This class inherits from \code{\link{matrix}}.
}
\seealso{
Other classes: 
\code{\link{CompositionMatrix-class}},
\code{\link{LogRatio-class}},
\code{\link{NumericMatrix-class}},
\code{\link{OutlierIndex-class}},
\code{\link{closure}()},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
