% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LogRatio-class}
\alias{LogRatio-class}
\alias{.LogRatio}
\alias{LR-class}
\alias{.LR}
\alias{CLR-class}
\alias{.CLR}
\alias{ALR-class}
\alias{.ALR}
\alias{ILR-class}
\alias{.ILR}
\alias{PLR-class}
\alias{.PLR}
\title{Log-Ratio Matrix}
\description{
S4 classes to represent log-ratio data transformations.
}
\section{Slots}{

\describe{
\item{\code{totals}}{A \code{\link{numeric}} vector to store the absolute row sums (before
the closure of the compositions).}

\item{\code{samples}}{A \code{\link{character}} vector to store the sample identifiers
(allows duplicates in case of repeated measurements).}

\item{\code{groups}}{A \code{\link{character}} vector to store the group names (if any).}

\item{\code{parts}}{A \code{\link{character}} vector to store the part names.}

\item{\code{ratio}}{A \code{\link{character}} vector to store the ratio names.}

\item{\code{order}}{An \code{\link{integer}} vector to store the original ordering of the
columns.}

\item{\code{base}}{A \code{\link{numeric}} matrix to store the basis of the transformation.}

\item{\code{weights}}{A \code{\link{numeric}} vector to store the weights assigned to the
respective log-ratios.}
}}

\note{
These classes inherit from \code{\link{matrix}}.
}
\seealso{
\code{\link[=transform_lr]{transform_lr()}}, \code{\link[=transform_clr]{transform_clr()}}, \code{\link[=transform_alr]{transform_alr()}},
\code{\link[=transform_ilr]{transform_ilr()}}, \code{\link[=transform_plr]{transform_plr()}}

Other classes: 
\code{\link{CompositionMatrix-class}},
\code{\link{LogicalMatrix-class}},
\code{\link{NumericMatrix-class}},
\code{\link{OutlierIndex-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
