% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LogRatio}
\alias{LogRatio}
\alias{LogRatio-class}
\alias{.LogRatio}
\alias{CLR-class}
\alias{.CLR}
\alias{ALR-class}
\alias{.ALR}
\alias{ILR-class}
\alias{.ILR}
\alias{PLR-class}
\alias{.PLR}
\title{Log-Ratio}
\description{
S4 classes to represent log-ratio data transformations.
}
\section{Slots}{

\describe{
\item{\code{parts}}{A \code{\link{character}} vector XXX.}

\item{\code{base}}{A \code{\link{numeric}} vector XXX.}

\item{\code{mean}}{A \code{\link{numeric}} vector XXX.}

\item{\code{denominator}}{An \code{\link{integer}} vector XXX.}

\item{\code{pivot}}{An \code{\link{integer}} vector XXX.}

\item{\code{ratio}}{A \code{\link{character}} vector XXX.}

\item{\code{norm}}{A \code{\link{numeric}} vector XXX.}
}}

\note{
These classes inherit from \code{\link{matrix}}.
}
\seealso{
Other log-ratio transformations: 
\code{\link{transform_alr}()},
\code{\link{transform_clr}()},
\code{\link{transform_ilr}()},
\code{\link{transform_inverse}()},
\code{\link{transform_plr}()}
}
\author{
N. Frerebeau
}
\concept{log-ratio transformations}
