% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{mahalanobis}
\alias{mahalanobis}
\alias{mahalanobis,CompositionMatrix-method}
\title{Mahalanobis Distance}
\usage{
\S4method{mahalanobis}{CompositionMatrix}(x, center, cov, robust = TRUE, ...)
}
\arguments{
\item{x}{A \linkS4class{CompositionMatrix} object.}

\item{center}{A \code{\link{numeric}} vector giving the mean vector of the
distribution. If missing, will be estimated from \code{x}.}

\item{cov}{A \code{\link{numeric}} matrix giving the covariance of the
distribution. If missing, will be estimated from \code{x}.}

\item{robust}{A \code{\link{logical}} scalar: should robust location and scatter
estimation be used (see \code{\link[robustbase:covMcd]{robustbase::covMcd()}})?}

\item{...}{Extra parameters to be passed to \code{\link[robustbase:covMcd]{robustbase::covMcd()}}.
Only used if \code{robust} is \code{TRUE}.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Computes the squared Mahalanobis distance of all rows in \code{x}.
}
\seealso{
\code{\link[stats:mahalanobis]{stats::mahalanobis()}}

Other statistics: 
\code{\link{covariance}},
\code{\link{dist}()},
\code{\link{variation}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
