% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutator}
\alias{mutator}
\alias{get}
\alias{set}
\alias{get_outliers}
\alias{get_outliers-method}
\alias{has_groups,OutlierIndex-method}
\alias{get_groups,OutlierIndex-method}
\alias{get_outliers,OutlierIndex-method}
\title{Get or Set Parts of an Object}
\usage{
get_outliers(x)

\S4method{has_groups}{OutlierIndex}(x)

\S4method{get_groups}{OutlierIndex}(x)

\S4method{get_outliers}{OutlierIndex}(x)
}
\arguments{
\item{x}{An object from which to get or set element(s) (typically a \verb{*Matrix}
object).}
}
\value{
An object of the same sort as \code{x} with the new values assigned.
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\seealso{
Other mutators: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
