% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{has_groups}
\alias{has_groups-method}
\alias{get_groups}
\alias{get_groups-method}
\alias{set_groups<-}
\alias{set_groups-method}
\alias{get_samples}
\alias{get_samples-method}
\alias{set_samples<-}
\alias{set_samples-method}
\alias{get_totals}
\alias{get_totals-method}
\alias{set_totals<-}
\alias{set_totals-method}
\alias{has_groups,CompositionMatrix-method}
\alias{get_groups,CompositionMatrix-method}
\alias{get_samples,CompositionMatrix-method}
\alias{get_totals,CompositionMatrix-method}
\alias{has_groups,LogRatio-method}
\alias{get_groups,LogRatio-method}
\alias{get_samples,LogRatio-method}
\alias{has_groups,OutlierIndex-method}
\alias{get_groups,OutlierIndex-method}
\alias{get_samples,OutlierIndex-method}
\alias{set_groups<-,CompositionMatrix-method}
\alias{set_groups,CompositionMatrix-method}
\alias{set_samples<-,CompositionMatrix-method}
\alias{set_samples,CompositionMatrix-method}
\alias{set_totals<-,CompositionMatrix-method}
\alias{set_totals,CompositionMatrix-method}
\title{Get or Set Parts of an Object}
\usage{
has_groups(x)

get_groups(x)

set_groups(x) <- value

get_samples(x)

set_samples(x) <- value

get_totals(x)

set_totals(x) <- value

\S4method{has_groups}{CompositionMatrix}(x)

\S4method{get_groups}{CompositionMatrix}(x)

\S4method{get_samples}{CompositionMatrix}(x)

\S4method{get_totals}{CompositionMatrix}(x)

\S4method{has_groups}{LogRatio}(x)

\S4method{get_groups}{LogRatio}(x)

\S4method{get_samples}{LogRatio}(x)

\S4method{has_groups}{OutlierIndex}(x)

\S4method{get_groups}{OutlierIndex}(x)

\S4method{get_samples}{OutlierIndex}(x)

\S4method{set_groups}{CompositionMatrix}(x) <- value

\S4method{set_samples}{CompositionMatrix}(x) <- value

\S4method{set_totals}{CompositionMatrix}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set element(s) (typically a \verb{*Matrix}
object).}

\item{value}{A possible value for the element(s) of \code{x}.}
}
\value{
\itemize{
\item \verb{set_*()} returns an object of the same sort as \code{x} with the new values
assigned.
\item \verb{get_*()} returns the part of \code{x}.
\item \verb{has_*()} returns a \code{\link{logical}} scalar.
}
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\details{
\describe{
\item{\code{get_samples(x)} and \code{get_samples(x) <- value}}{Get or set
the sample names of \code{x}.}
\item{\code{get_groups(x)} and \code{set_groups(x) <- value}}{Get or set
the groups of \code{x}.}
}
}
\seealso{
Other mutators: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
