% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexus-package.R
\docType{package}
\name{nexus-package}
\alias{nexus-package}
\alias{nexus}
\title{nexus: Sourcing Archaeological Materials by Chemical Composition}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Exploration and analysis of compositional data in the framework of Aitchison (1986, ISBN: 978-94-010-8324-9). This package provides tools for chemical fingerprinting and source tracking of ancient materials.
}
\details{
\tabular{ll}{
\strong{Version} \tab 0.5.0 \cr
\strong{License} \tab GPL-3 \cr
\strong{CRAN DOI} \tab \doi{10.32614/CRAN.package.nexus} \cr
\strong{Zenodo DOI} \tab \doi{10.5281/zenodo.10225630} \cr
}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\section{Package options}{

\pkg{nexus} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{nexus.verbose}: a \code{\link{logical}} scalar. Should \R report extra information
on progress? Defaults to \code{\link[=interactive]{interactive()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/tesselle/nexus}
  \item \url{https://packages.tesselle.org/nexus/}
  \item \url{https://tesselle.r-universe.dev/nexus}
  \item Report bugs at \url{https://codeberg.org/tesselle/nexus/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})

Authors:
\itemize{
  \item Anne Philippe \email{anne.philippe@univ-nantes.fr} (\href{https://orcid.org/0000-0002-5331-5087}{ORCID})
}

Other contributors:
\itemize{
  \item Brice Lebrun (\href{https://orcid.org/0000-0001-7503-8685}{ORCID}) (Logo designer) [artist]
  \item Université Bordeaux Montaigne (03pbgwk21) [funder]
  \item CNRS (02feahw73) [funder]
}

}
\keyword{internal}
