% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NumericMatrix-class}
\alias{NumericMatrix-class}
\alias{.NumericMatrix}
\title{Numeric Matrix}
\description{
S4 classes that represent a \eqn{m \times p}{m x p} numeric matrix.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A \eqn{m \times p}{m x p} \code{numeric} \code{\link{matrix}}.}
}}

\note{
This class inherits from \code{\link{matrix}}.
}
\seealso{
Other classes: 
\code{\link{CompositionMatrix-class}},
\code{\link{GroupedComposition-class}},
\code{\link{GroupedLogRatio-class}},
\code{\link{LogRatio-class}},
\code{\link{OutlierIndex-class}},
\code{\link{ReferenceGroups-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
