% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NumericMatrix-class}
\alias{NumericMatrix-class}
\alias{.NumericMatrix}
\title{Numeric Matrix}
\description{
S4 classes that represent a \eqn{m \times p}{m x p} numeric matrix.
}
\note{
This class inherits from \code{\link{matrix}}.
}
\seealso{
Other classes: 
\code{\link{CompositionMatrix-class}},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
