% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/outliers.R
\docType{methods}
\name{outliers}
\alias{outliers}
\alias{outliers-method}
\alias{outliers,CompositionMatrix-method}
\title{Outlier Detection}
\usage{
outliers(object, ...)

\S4method{outliers}{CompositionMatrix}(
  object,
  ...,
  groups = get_groups(object),
  method = c("mve", "mcd"),
  quantile = 0.975
)
}
\arguments{
\item{object}{A \code{\linkS4class{CompositionMatrix}}.}

\item{...}{Extra parameters to be passed to \code{\link[MASS:cov.rob]{MASS::cov.rob()}}.
Only used if \code{robust} is \code{TRUE}.}

\item{groups}{A \code{vector} of grouping elements, as long as the variables in \code{object}.}

\item{method}{A \code{\link{character}} string specifying the method to be used.
It must be one of "\code{mve}" (minimum volume ellipsoid) or "\code{mcd}" (minimum
covariance determinant). Only used if \code{robust} is \code{TRUE}.}

\item{quantile}{A length-one \code{\link{numeric}} vector giving the significance level.
\code{quantile} is used as a cut-off value for outlier detection: observations
with larger (squared) Mahalanobis distance are considered as potential
outliers.}
}
\value{
An \code{\linkS4class{OutlierIndex}} object.
}
\description{
Outlier Detection
}
\details{
An outlier can be defined as having a very large Mahalanobis distance from
all observations. In this way, a certain proportion of the observations can
be identified, e.g. the top 2\% of values (i.e. values above the 0.98th
percentile of the Chi-2 distribution).

On the one hand, the Mahalanobis distance is likely to be strongly
affected by the presence of outliers. Rousseeuw and van Zomeren (1990) thus
recommend using robust methods (which are not excessively affected by the
presence of outliers).

On the other hand, the choice of the threshold for classifying an
observation as an outlier should be discussed. There is no apparent reason
why a particular threshold should be applicable to all data sets
(Filzmoser, Garrett, and Reimann 2005).
}
\examples{
## Data from Day et al. 2011
data("kommos", package = "folio") # Coerce to compositional data
kommos <- remove_NA(kommos, margin = 1) # Remove cases with missing values
coda <- as_composition(kommos, groups = 1) # Use ceramic types for grouping

## Detect outliers
out <- outliers(coda, groups = NULL, method = "mcd")

plot(out, type = "dotchart")
plot(out, type = "distance")
plot(out, type = "qqplot")

## Detect outliers by group
out <- outliers(coda[, 1:15], method = "mcd")

plot(out, type = "dotchart", select = 1, robust = FALSE)
plot(out, type = "dotchart", select = 2, robust = FALSE)
plot(out, type = "dotchart", select = 3, robust = FALSE)
plot(out, type = "dotchart", select = 4, robust = FALSE)
}
\references{
Filzmoser, P., Garrett, R. G. & Reimann, C. (2005). Multivariate outlier
detection in exploration geochemistry. \emph{Computers & Geosciences},
31(5), 579-587. \doi{10.1016/j.cageo.2004.11.013}.

Filzmoser, P. & Hron, K. (2008). Outlier Detection for Compositional Data
Using Robust Methods. \emph{Mathematical Geosciences}, 40(3), 233-248.
\doi{10.1007/s11004-007-9141-5}.

Filzmoser, P., Hron, K. & Reimann, C. (2012). Interpretation of multivariate
outliers for compositional data. \emph{Computers & Geosciences}, 39, 77-85.
\doi{10.1016/j.cageo.2011.06.014}.

Rousseeuw, P. J. & van Zomeren, B. C. (1990). Unmasking Multivariate Outliers
and Leverage Points. \emph{Journal of the American Statistical Association},
85(411): 633-639. \doi{10.1080/01621459.1990.10474920}.

Santos, F. (2020). Modern methods for old data: An overview of some robust
methods for outliers detection with applications in osteology. \emph{Journal of
Archaeological Science: Reports}, 32, 102423.
\doi{10.1016/j.jasrep.2020.102423}.
}
\seealso{
Other outlier detection methods: 
\code{\link{plot_outliers}}
}
\author{
N. Frerebeau
}
\concept{outlier detection methods}
