% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/pca.R
\docType{methods}
\name{pca_coda}
\alias{pca_coda}
\alias{pca,CompositionMatrix-method}
\alias{pca,LogRatio-method}
\title{Principal Components Analysis}
\usage{
\S4method{pca}{CompositionMatrix}(
  object,
  center = TRUE,
  scale = TRUE,
  rank = NULL,
  sup_row = NULL,
  sup_col = NULL,
  weight_row = NULL,
  weight_col = NULL
)

\S4method{pca}{LogRatio}(
  object,
  center = TRUE,
  scale = TRUE,
  rank = NULL,
  sup_row = NULL,
  sup_col = NULL,
  weight_row = NULL,
  weight_col = NULL
)
}
\arguments{
\item{object}{A \linkS4class{LogRatio} object.}

\item{center}{A \code{\link{logical}} scalar: should the variables be shifted to be
zero centered?}

\item{scale}{A \code{\link{logical}} scalar: should the variables be scaled to unit
variance?}

\item{rank}{An \code{\link{integer}} value specifying the maximal number of components
to be kept in the results. If \code{NULL} (the default), \eqn{p - 1} components
will be returned.}

\item{sup_row}{A \code{\link{numeric}} or \code{\link{logical}} vector specifying the indices of
the supplementary rows (individuals).}

\item{sup_col}{A \code{\link{numeric}} or \code{\link{logical}} vector specifying the indices of
the supplementary columns (variables).}

\item{weight_row}{A \code{\link{numeric}} vector specifying the active row (individual)
weights. If \code{NULL} (the default), uniform weights are used. Row weights are
internally normalized to sum 1}

\item{weight_col}{A \code{\link{numeric}} vector specifying the active column
(variable) weights. If \code{NULL} (the default), uniform weights (1) are
used.}
}
\value{
A \link[dimensio:PCA-class]{dimensio::PCA} object.
}
\description{
Computes a principal components analysis based on the singular value
decomposition.
}
\examples{
## Coerce to compositional data
data("hongite")
coda <- as_composition(hongite)

## Centered log-ratio
clr <- transform_clr(coda)

## PCA
X <- pca(clr)

## Plot
viz_individuals(X)
viz_variables(X)
}
\seealso{
\code{\link[dimensio:pca]{dimensio::pca()}}, \code{\link[dimensio:biplot]{dimensio::biplot()}}
}
\author{
N. Frerebeau
}
\concept{multivariate analysis}
