% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{petrography}
\alias{petrography}
\title{Can Sora Petrographic Data}
\format{
A \code{\link{data.frame}} with 21 variables:
\describe{
\item{VAR1}{Optical activity.}
\item{VAR2}{Inclusion orientation.}
\item{VAR3}{Void orientation.}
\item{VAR4}{Texture.}
\item{VAR5}{Special components.}
\item{VAR6}{Plutonic rocks.}
\item{VAR7}{Volcanic rocks.}
\item{VAR8}{Metamorphic rocks.}
\item{VAR9}{Sedimentary rocks.}
\item{VAR10}{Quartz.}
\item{VAR11}{Feldspar.}
\item{VAR12}{Plagioclase.}
\item{VAR13}{Pyroxenes.}
\item{VAR14}{Amphiboles.}
\item{VAR15}{Micas.}
\item{VAR16}{Phyllosilicates.}
\item{VAR17}{Carbonates.}
\item{VAR18}{Other constituents.}
\item{VAR19}{Packing.}
}
}
\usage{
petrography
}
\description{
Can Sora Petrographic Data
}
\references{
Cau, M.-A., Day, P. M., Baxter, M. J., Papageorgiou, I., Iliopoulos, I. &
Montana, G. (2004). Exploring Automatic Grouping Procedures in Ceramic
Petrology. \emph{Journal of Archaeological Science}, 31(9): 1325-1338.
\doi{10.1016/j.jas.2004.03.006}.
}
\seealso{
Other datasets: 
\code{\link{arctic}},
\code{\link{chemistry}},
\code{\link{hongite}},
\code{\link{slides}}
}
\concept{datasets}
\keyword{datasets}
