% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/transform.R
\docType{methods}
\name{transform_pivot}
\alias{transform_pivot}
\alias{transform_pivot-method}
\alias{transform_pivot,CompositionMatrix-method}
\title{Pivot Log-Ratios (PLR)}
\usage{
transform_pivot(object, ...)

\S4method{transform_pivot}{CompositionMatrix}(object, ordering = seq_len(ncol(object)), weights = FALSE)
}
\arguments{
\item{object}{A \linkS4class{CompositionMatrix} object.}

\item{...}{Currently not used.}

\item{ordering}{An \code{\link{integer}} vector specifying the ordering of the parts to
be used.}

\item{weights}{A \code{\link{logical}} scalar: sould a varying weight be used. If
\code{FALSE} (the default), equally-weighted parts are used. Alternatively, a
positive \code{\link{numeric}} vector of weights can be specified.}
}
\value{
A \linkS4class{PLR} object.
}
\description{
Computes PLR transformations.
}
\examples{
## Coerce to chemical data
data("hongite")
coda <- as_composition(hongite)

## Centered log-ratio
clr <- transform_clr(coda)

## Additive log-ratio
alr <- transform_alr(coda)

## Isometric log-ratio
ilr1 <- transform_ilr(coda)
ilr2 <- transform_pivot(coda)
}
\references{
Fišerová, E. & Hron, K. (2011). On the Interpretation of Orthonormal
Coordinates for Compositional Data. \emph{Mathematical Geosciences}, 43(4),
455‑468. \doi{10.1007/s11004-011-9333-x}.

Hron, K., Filzmoser, P., de Caritat, P., Fišerová, E. & Gardlo, A. (2017).
Weighted Pivot Coordinates for Compositional Data and Their Application to
Geochemical Mapping. \emph{Mathematical Geosciences}, 49(6), 797-814.
\doi{10.1007/s11004-017-9684-z}.

Greenacre, M. J. (2019). \emph{Compositional Data Analysis in Practice}.
Boca Raton: CRC Press.
}
\seealso{
Other log-ratio transformations: 
\code{\link{LogRatio}},
\code{\link{transform_alr}()},
\code{\link{transform_clr}()},
\code{\link{transform_ilr}()},
\code{\link{transform_inverse}()}
}
\author{
N. Frerebeau
}
\concept{log-ratio transformations}
