% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predator}
\alias{predator}
\title{Predator-Prey Compositions}
\format{
A \code{\link{data.frame}} with 3 variables (proportions):
\describe{
\item{\code{P}}{Predator.}
\item{\code{Q}}{Prey of species Q.}
\item{\code{R}}{Prey of species R.}
}
}
\source{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall. \doi{10.1007/978-94-009-4109-0}.
}
\usage{
predator
}
\description{
Predator-prey compositions at 25 different sites.
}
\seealso{
Other datasets: 
\code{\link{arctic}},
\code{\link{boxite}},
\code{\link{coxite}},
\code{\link{hongite}},
\code{\link{kongite}},
\code{\link{lava}},
\code{\link{slides}}
}
\concept{datasets}
\keyword{datasets}
