% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{replace_NA}
\alias{replace_zero}
\alias{remove_NA}
\alias{remove_zero}
\alias{pca}
\alias{biplot}
\alias{screeplot}
\alias{viz_individuals}
\alias{viz_variables}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{arkhe}{\code{\link[arkhe:missing]{remove_NA}}, \code{\link[arkhe:zero]{remove_zero}}, \code{\link[arkhe:missing]{replace_NA}}, \code{\link[arkhe:zero]{replace_zero}}}

  \item{dimensio}{\code{\link[dimensio]{biplot}}, \code{\link[dimensio]{pca}}, \code{\link[dimensio]{screeplot}}, \code{\link[dimensio]{viz_individuals}}, \code{\link[dimensio]{viz_variables}}}
}}

