% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slides}
\alias{slides}
\title{Thin Sections Compositions}
\format{
A \code{\link{data.frame}} with 9 variables:
\describe{
\item{\code{analyst}}{Analyst number.}
\item{\code{slide}}{Slide number.}
\item{\code{quartz}}{Quartz (percent).}
\item{\code{microcline}}{Microcline (percent).}
\item{\code{plagioclass}}{Plagioclass (percent).}
\item{\code{biotite}}{Biotite (percent).}
\item{\code{muscovite}}{Muscovite (percent).}
\item{\code{opaques}}{Opaque minerals (percent).}
\item{\code{nonopaques}}{Non-opaque minerals (percent).}
}
}
\usage{
slides
}
\description{
Mineral compositions of five slides as reported by five analysts.
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall.
}
\seealso{
Other datasets: 
\code{\link{arctic}},
\code{\link{boxite}},
\code{\link{coxite}},
\code{\link{hongite}},
\code{\link{kongite}},
\code{\link{lava}},
\code{\link{predator}}
}
\concept{datasets}
\keyword{datasets}
