% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{spread}
\alias{spread}
\alias{mvar}
\alias{mvar-method}
\alias{msd}
\alias{msd-method}
\alias{mvar,CompositionMatrix-method}
\alias{msd,CompositionMatrix-method}
\title{Metric Variance and Standard Deviation}
\usage{
mvar(x, ...)

msd(x, ...)

\S4method{mvar}{CompositionMatrix}(x)

\S4method{msd}{CompositionMatrix}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{CompositionMatrix}} object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
\itemize{
\item \code{mvar()} computes the metric variance, i.e. a global measure of spread.
\item \code{msd()} computes the metric standard deviation.
}
}
\examples{
## Coerce to compositional data
data("hongite")
coda <- as_composition(hongite)

## Mean
mean(coda)

## Metric variance
mvar(coda)

## Metric standard deviation
msd(coda)
}
\references{
Pawlowsky-Glahn, V. and Egozcue, J. J. (2001). Geometric Approach to
Statistical Analysis on the Simplex. \emph{Stochastic Environmental Research and
Risk Assessment}, 15(5): 384-398. \doi{10.1007/s004770100077}.
}
\seealso{
Other statistics: 
\code{\link{aggregate}()},
\code{\link{covariance}()},
\code{\link{dist}()},
\code{\link{mahalanobis}()},
\code{\link{mean}()},
\code{\link{variation}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
