% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{totals}
\alias{totals}
\alias{total}
\alias{total-method}
\alias{total<-}
\alias{total,CompositionMatrix-method}
\alias{total,LogRatio-method}
\alias{total<-,CompositionMatrix-method}
\title{Row Sums}
\usage{
total(x)

total(x) <- value

\S4method{total}{CompositionMatrix}(x)

\S4method{total}{LogRatio}(x)

\S4method{total}{CompositionMatrix}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set \code{totals}.}

\item{value}{A possible value for the \code{totals} of \code{x}.}
}
\value{
\itemize{
\item \verb{total()<-} returns an object of the same sort as \code{x} with the new
row sums assigned.
\item \code{total()} returns the row sums of \code{x}.
}
}
\description{
Retrieves or defines the row sums (before closure).
}
\examples{
## Create a count matrix
A1 <- matrix(data = sample(1:100, 100, TRUE), nrow = 20)

## Coerce to compositions
B <- as_composition(A1)

## Row sums are internally stored before coercing to relative frequencies
total(B)

## This allows to restore the source data
A2 <- as_amounts(B)

## Coerce to a data.frame
X <- data.frame(B)
head(X)
}
\seealso{
Other mutators: 
\code{\link{groups}},
\code{\link{mutators}}
}
\author{
N. Frerebeau
}
\concept{mutators}
