% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/transform.R
\docType{methods}
\name{transform}
\alias{transform}
\alias{transform_clr}
\alias{transform_clr-method}
\alias{transform_alr}
\alias{transform_alr-method}
\alias{transform_ilr}
\alias{transform_ilr-method}
\alias{transform_pivot}
\alias{transform_pivot-method}
\alias{transform_clr,CompositionMatrix-method}
\alias{transform_alr,CompositionMatrix-method}
\alias{transform_ilr,CompositionMatrix-method}
\alias{transform_pivot,CompositionMatrix-method}
\title{Data Transformation}
\usage{
transform_clr(object, ...)

transform_alr(object, ...)

transform_ilr(object, ...)

transform_pivot(object, ...)

\S4method{transform_clr}{CompositionMatrix}(object, base = exp(1))

\S4method{transform_alr}{CompositionMatrix}(object, j = ncol(object), base = exp(1))

\S4method{transform_ilr}{CompositionMatrix}(object, base = exp(1))

\S4method{transform_pivot}{CompositionMatrix}(object, pivot = 1, base = exp(1))
}
\arguments{
\item{object}{A \linkS4class{ChemicalMatrix} object.}

\item{...}{Currently not used.}

\item{base}{A positive \code{\link{numeric}} value giving the base with
respect to which logarithms are computed. Change this only if you know
what your are doing.}

\item{j}{An \code{\link{integer}} giving the index of the rationing part.}

\item{pivot}{An \code{\link{integer}} giving the index of the pivotal
variable.}
}
\value{
A \linkS4class{LogRatio} object.
}
\description{
Data Transformation
}
\examples{
## Coerce to chemical data
data("kommos", package = "folio")
coda <- as_composition(kommos[, -c(1, 2)])
coda <- remove_NA(coda, margin = 1)

## Centered log-ratio
clr <- transform_clr(coda)

## Additive log-ratio
alr <- transform_alr(coda, j = "Ca")

## Isometric log-ratio
ilr1 <- transform_ilr(coda)
ilr2 <- transform_pivot(coda)
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall.
DOI: \href{https://dx.doi.org/10.1007/978-94-009-4109-0}{10.1007/978-94-009-4109-0}.

Egozcue, J. J., Pawlowsky-Glahn, V., Mateu-Figueras, G. & Barceló-Vidal, C.
(2003). Isometric Logratio Transformations for Compositional Data Analysis.
\emph{Mathematical Geology}, 35(3), 279-300.
DOI: \href{https://dx.doi.org/10.1023/A:1023818214614}{10.1023/A:1023818214614}.

Fišerová, E. & Hron, K. (2011). On the Interpretation of Orthonormal
Coordinates for Compositional Data. \emph{Mathematical Geosciences}, 43(4),
455‑468. DOI:
\href{https://doi.org/10.1007/s11004-011-9333-x}{10.1007/s11004-011-9333-x}.
}
\author{
N. Frerebeau
}
\concept{transformation}
