% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/transform.R
\docType{methods}
\name{transform_ilr}
\alias{transform_ilr}
\alias{transform_ilr-method}
\alias{transform_ilr,CompositionMatrix-method}
\title{Isometric Log-Ratios (ILR)}
\usage{
transform_ilr(object, ...)

\S4method{transform_ilr}{CompositionMatrix}(object)
}
\arguments{
\item{object}{A \linkS4class{CompositionMatrix} object.}

\item{...}{Currently not used.}
}
\value{
An \linkS4class{ILR} object.
}
\description{
Computes ILR transformations.
}
\details{
The ILR transformation provides the coordinates of any composition with
respect to a given orthonormal basis. \code{transform_ilr()} uses the orthonormal
basis (Helmert matrix) originally defined by Egozcue \emph{et al.} (2003).
}
\examples{
## Coerce to compositonal data
data("hongite")
coda <- as_composition(hongite)

## Pairwise log-ratio
lr <- transform_lr(coda)

## Centered log-ratio
clr <- transform_clr(coda)

## Additive log-ratio
alr <- transform_alr(coda)

## Isometric log-ratio
ilr1 <- transform_ilr(coda)
ilr2 <- transform_plr(coda)
}
\references{
Egozcue, J. J., Pawlowsky-Glahn, V., Mateu-Figueras, G. & Barceló-Vidal, C.
(2003). Isometric Logratio Transformations for Compositional Data Analysis.
\emph{Mathematical Geology}, 35(3), 279-300. \doi{10.1023/A:1023818214614}.

Greenacre, M. J. (2019). \emph{Compositional Data Analysis in Practice}.
Boca Raton: CRC Press.
}
\seealso{
Other log-ratio transformations: 
\code{\link{LogRatio}},
\code{\link{transform_alr}()},
\code{\link{transform_clr}()},
\code{\link{transform_inverse}()},
\code{\link{transform_lr}()},
\code{\link{transform_plr}()}
}
\author{
N. Frerebeau
}
\concept{log-ratio transformations}
