% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/transform.R
\docType{methods}
\name{transform_plr}
\alias{transform_plr}
\alias{transform_plr-method}
\alias{transform_plr,CompositionMatrix-method}
\title{Pivot Log-Ratios (PLR)}
\usage{
transform_plr(object, ...)

\S4method{transform_plr}{CompositionMatrix}(object, pivot = 1)
}
\arguments{
\item{object}{A \linkS4class{CompositionMatrix} object.}

\item{...}{Currently not used.}

\item{pivot}{An \code{\link{integer}} giving the index of the pivotal variable.}
}
\value{
A \linkS4class{PLR} object.
}
\description{
Computes PLR transformations.
}
\examples{
## Coerce to chemical data
data("hongite")
coda <- as_composition(hongite)

## Centered log-ratio
clr <- transform_clr(coda)

## Additive log-ratio
alr <- transform_alr(coda)

## Isometric log-ratio
ilr1 <- transform_ilr(coda)
ilr2 <- transform_plr(coda)
}
\references{
Fišerová, E. & Hron, K. (2011). On the Interpretation of Orthonormal
Coordinates for Compositional Data. \emph{Mathematical Geosciences}, 43(4),
455‑468. \doi{10.1007/s11004-011-9333-x}.

Hron, K., Filzmoser, P., de Caritat, P., Fišerová, E. & Gardlo, A. (2017).
Weighted Pivot Coordinates for Compositional Data and Their Application to
Geochemical Mapping. \emph{Mathematical Geosciences}, 49(6), 797-814.
\doi{10.1007/s11004-017-9684-z}.

Greenacre, M. J. (2019). \emph{Compositional Data Analysis in Practice}.
Boca Raton: CRC Press.
}
\seealso{
Other log-ratio transformations: 
\code{\link{LogRatio}},
\code{\link{transform_alr}()},
\code{\link{transform_clr}()},
\code{\link{transform_ilr}()},
\code{\link{transform_inverse}()}
}
\author{
N. Frerebeau
}
\concept{log-ratio transformations}
