% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{weights}
\alias{weights}
\alias{get}
\alias{set}
\alias{weights,ALR-method}
\alias{weights,LR-method}
\alias{weights,LogRatio-method}
\title{Weights}
\usage{
\S4method{weights}{ALR}(object, ...)

\S4method{weights}{LR}(object, ...)

\S4method{weights}{LogRatio}(object, ...)
}
\arguments{
\item{object}{An object for which the extraction of weights is meaningful.}

\item{...}{Currently not used.}
}
\description{
Extract weights from an object.
}
\seealso{
Other mutators: 
\code{\link{as.data.frame}()},
\code{\link{labels}()},
\code{\link{totals}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
