% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemistry}
\alias{chemistry}
\title{Can Sora Chemical Data}
\format{
A \code{\link{data.frame}} with 30 variables.
}
\usage{
chemistry
}
\description{
Can Sora Chemical Data
}
\references{
Cau, M.-A. (1999). Importaciones de cerámica tardorromana de cocina en las
Iles Balears: el caso de Can Sora (Eivissa). In J. Capel Martinez,
\emph{Arqueometria y Arqueologia}, p. 197-219. Granada: Editorial Universidad de
Granada. Monografica Arte y Arqueología 47.
}
\seealso{
Other datasets: 
\code{\link{hongite}},
\code{\link{petrography}}
}
\concept{datasets}
\keyword{datasets}
