% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/simplex.R
\docType{methods}
\name{closure}
\alias{closure}
\alias{closure-method}
\alias{closure,numeric-method}
\alias{closure,matrix-method}
\title{Closure Operation}
\usage{
closure(x, ...)

\S4method{closure}{numeric}(x, total = 1, na.rm = FALSE)

\S4method{closure}{matrix}(x, total = 1, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector or matrix.}

\item{...}{Currently not used.}

\item{total}{A \link{numeric} vector specifying the total amount to which the
compositions should be closed (defaults to 1).}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values be removed?}
}
\value{
A \code{\link{numeric}} vector or matrix (same as \code{x}).
}
\description{
Closes compositions to sum up to 1.
}
\examples{
## Create a count matrix
A1 <- matrix(data = as.numeric(sample(1:100, 100, TRUE)), nrow = 20)

## Coerce to compositions
B <- as_composition(A1)

## Row sums are internally stored before coercing to relative frequencies
## (use get_totals() to retrieve these values)
## This allows to restore the source data
A2 <- as_amounts(B)

## Coerce to an S3 data.frame
X <- data.frame(B)
head(X)
}
\seealso{
Other operations in the simplex: 
\code{\link{arithmetic}},
\code{\link{perturbation}()},
\code{\link{powering}()},
\code{\link{scalar}()}
}
\author{
N. Frerebeau
}
\concept{operations in the simplex}
