% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R, R/mutators.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{as_count}
\alias{as_count-method}
\alias{as_composition}
\alias{as_composition-method}
\alias{get_totals}
\alias{get_totals-method}
\alias{set_totals<-}
\alias{set_totals-method}
\alias{as_composition,matrix-method}
\alias{as_composition,data.frame-method}
\alias{as_count,CompositionMatrix-method}
\alias{get_totals,CompositionMatrix-method}
\alias{set_totals<-,CompositionMatrix-method}
\alias{set_totals,CompositionMatrix-method}
\title{Coerce}
\usage{
as_count(from, ...)

as_composition(from, ...)

get_totals(x)

set_totals(x) <- value

\S4method{as_composition}{matrix}(from)

\S4method{as_composition}{data.frame}(from, samples = NULL, groups = NULL)

\S4method{as_count}{CompositionMatrix}(from)

\S4method{get_totals}{CompositionMatrix}(x)

\S4method{set_totals}{CompositionMatrix}(x) <- value
}
\arguments{
\item{from}{An object to be coerced.}

\item{...}{Currently not used.}

\item{x}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{x}.}

\item{samples}{An \code{\link{integer}} giving the index of the column to be used for
sample identification: allows to identify replicated measurements.
If \code{NULL} (the default), row names will be used as sample IDs.}

\item{groups}{An \code{\link{integer}} giving the index of the column to be used to
group the samples. If \code{NULL} (the default), no grouping is stored.}
}
\value{
A coerced object.
}
\description{
Coerces an object to a \code{CompositionMatrix} object.
}
\details{
The following methods are available:

\tabular{lll}{
\strong{Method} \tab \strong{From} \tab \strong{To} \cr
\code{as_composition()} \tab \code{\link{matrix}} or \code{\link{data.frame}} \tab \linkS4class{CompositionMatrix} \cr
\code{as_count()} \tab \linkS4class{CompositionMatrix} \tab \code{\link{matrix}} \cr
}

Note that all non-numeric variable will be removed.

The \linkS4class{CompositionMatrix} class has special slots:
\itemize{
\item \code{samples} for replicated measurements/observation,
\item \code{groups} to group data by site/area.
}

When coercing a \code{data.frame} to a \linkS4class{CompositionMatrix} object, an
attempt is made to automatically assign values to these slots by mapping
column names (case insensitive, plural insensitive). This behavior can be
disabled by setting \code{options(nexus.autodetect = FALSE)} or overridden by
explicitly specifying the columns to be used in \code{as_composition()}.
}
\examples{
## Create a count matrix
A1 <- matrix(data = as.numeric(sample(1:100, 100, TRUE)), nrow = 20)

## Coerce to compositions
B <- as_composition(A1)

## Row sums are internally stored before coercing to relative frequencies
## (use get_totals() to retrieve these values)
## This allows to restore the source data
A2 <- as_count(B)

## Coerce to an S3 data.frame
X <- data.frame(B)
head(X)
}
\seealso{
Other classes: 
\code{\link{CompositionMatrix-class}},
\code{\link{LogRatio-class}},
\code{\link{NumericMatrix-class}},
\code{\link{OutlierIndex-class}}
\code{\link{closure}()}
}
\author{
N. Frerebeau
}
\concept{classes}
