% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/condense.R
\docType{methods}
\name{condense}
\alias{condense}
\alias{condense-method}
\alias{condense,CompositionMatrix-method}
\title{Compositional Mean of Data Subsets}
\usage{
condense(x, ...)

\S4method{condense}{CompositionMatrix}(x, by = get_samples(x), na.rm = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{CompositionMatrix}} object.}

\item{...}{Currently not used.}

\item{by}{A \code{vector} of grouping elements, as long as the variables in \code{x}.}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values be removed?}
}
\value{
A \code{\linkS4class{CompositionMatrix}} object.
}
\description{
Splits the data into subsets and computes compositional mean for each.
}
\examples{
## Data from Aitchison 1986
data("slides")

## Coerce to a compositional matrix
coda <- as_composition(slides, sample = 2, group = 1)

## Compositional mean by sample
condense(coda, by = get_samples(coda))

## Compositional mean by group
condense(coda, by = get_groups(coda))
}
\seealso{
\code{\link[=mean]{mean()}}, \code{\link[=aggregate]{aggregate()}}

Other statistics: 
\code{\link{aggregate}()},
\code{\link{covariance}()},
\code{\link{dist}},
\code{\link{mahalanobis}()},
\code{\link{margin}()},
\code{\link{mean}()},
\code{\link{metric_var}()},
\code{\link{quantile}()},
\code{\link{scale}()},
\code{\link{variation}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
