% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{groups}
\alias{groups}
\alias{get_groups}
\alias{get_groups-method}
\alias{get_groups,OutlierIndex-method}
\title{Working With Groups}
\usage{
get_groups(x)

\S4method{get_groups}{OutlierIndex}(x)
}
\arguments{
\item{x}{An object from which to get or set \code{groups}.}
}
\value{
\itemize{
\item \code{get_groups()} returns a \code{\link{character}} vector giving the group names of \code{x}.
}
}
\description{
Retrieves or defines the groups to which the observations belong.
}
\examples{
## Data from Aitchison 1986
data("slides")
head(slides)

## Coerce to compositional data
coda <- as_composition(slides)
head(augment(coda))
}
\seealso{
Other mutators: 
\code{\link{split}()},
\code{\link{subset}()},
\code{\link{totals}}
}
\author{
N. Frerebeau
}
\concept{mutators}
