% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{groups}
\alias{groups}
\alias{has_groups}
\alias{has_groups-method}
\alias{get_groups}
\alias{get_groups-method}
\alias{set_groups<-}
\alias{set_groups-method}
\alias{has_groups,CompositionMatrix-method}
\alias{has_groups,LogRatio-method}
\alias{has_groups,OutlierIndex-method}
\alias{get_groups,CompositionMatrix-method}
\alias{get_groups,LogRatio-method}
\alias{get_groups,OutlierIndex-method}
\alias{set_groups<-,CompositionMatrix-method}
\alias{set_groups,CompositionMatrix-method}
\title{Deal With Groups}
\usage{
has_groups(x)

get_groups(x)

set_groups(x) <- value

\S4method{has_groups}{CompositionMatrix}(x)

\S4method{has_groups}{LogRatio}(x)

\S4method{has_groups}{OutlierIndex}(x)

\S4method{get_groups}{CompositionMatrix}(x)

\S4method{get_groups}{LogRatio}(x)

\S4method{get_groups}{OutlierIndex}(x)

\S4method{set_groups}{CompositionMatrix}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set \code{groups}.}

\item{value}{A possible value for the \code{groups} of \code{x}.}
}
\value{
\itemize{
\item \code{set_groups()} returns an object of the same sort as \code{x} with the new
group names assigned.
\item \code{get_groups()} returns the group names of \code{x}.
\item \code{has_groups()} returns a \code{\link{logical}} scalar.
}
}
\description{
Retrieves or defines the groups to which the observations belong.
}
\details{
See \code{vignette("manual")}.
}
\seealso{
Other mutators: 
\code{\link{samples}},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
