% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hongite}
\alias{hongite}
\title{Hongite Mineralogy}
\format{
A \code{\link{data.frame}} with 5 variables (minerals):
\describe{
\item{A}{Albite (percent).}
\item{B}{Blandite (percent).}
\item{C}{Cornite (percent).}
\item{D}{Daubite (percent).}
\item{E}{Endite (percent).}
}
}
\usage{
hongite
}
\description{
Mineral compositions of 25 rock specimens of hongite type.
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall.
}
\seealso{
Other datasets: 
\code{\link{arctic}},
\code{\link{slides}}
}
\concept{datasets}
\keyword{datasets}
