% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{mean}
\alias{mean}
\alias{mean,CompositionMatrix-method}
\title{Compositional Mean}
\usage{
\S4method{mean}{CompositionMatrix}(x, ...)
}
\arguments{
\item{x}{A \linkS4class{CompositionMatrix} object.}

\item{...}{Further argument to be passed to other methods.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Compositional Mean
}
\details{
Closed vector of the columns geometric means.
}
\examples{
## Coerce to compositional data
data("hongite")
coda <- as_composition(hongite)

## Mean
mean(coda)
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall, p. 64-91. \doi{10.1007/978-94-009-4109-0}.
}
\seealso{
Other statistics: 
\code{\link{covariance}},
\code{\link{dist}()},
\code{\link{mahalanobis}()},
\code{\link{variation}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
