% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{labels,CompositionMatrix-method}
\alias{labels,LogRatio-method}
\alias{weights,LogRatio-method}
\title{Get or Set Parts of an Object}
\usage{
\S4method{labels}{CompositionMatrix}(object, ...)

\S4method{labels}{LogRatio}(object, ...)

\S4method{weights}{LogRatio}(object, ...)
}
\arguments{
\item{object}{An object from which to get or set element(s).}

\item{...}{Currently not used.}
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\seealso{
Other mutators: 
\code{\link{groups}},
\code{\link{totals}}
}
\author{
N. Frerebeau
}
\concept{mutators}
