% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexus-package.R
\docType{package}
\name{nexus-package}
\alias{nexus-package}
\alias{_PACKAGE}
\alias{nexus}
\title{nexus: Sourcing Archaeological Materials by Chemical Composition}
\description{
Sourcing Archaeological Materials by Chemical Composition.
}
\details{
\tabular{ll}{
\strong{Package:} \tab nexus \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0.9000 \cr
\strong{License:} \tab GPL-3 \cr
\strong{DOI:} \tab \doi{xxx}
}
}
\note{
This work received a state financial support managed by the Agence Nationale
de la Recherche (France) throught the programs
\emph{Investissements d'avenir}
(ref. \href{https://lascarbx.labex.u-bordeaux.fr}{ANR-10-LABX-52}) and
\emph{IRANGKOR} (ref.
\href{https://anr.fr/Project-ANR-14-CE31-0007}{ANR-14-CE31-0007}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nexus.tesselle.org}
  \item \url{https://github.com/tesselle/nexus,}
  \item Report bugs at \url{https://github.com/tesselle/nexus/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
Anne Philippe \tab \emph{Université de Nantes, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
