% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexus-package.R
\docType{package}
\name{nexus-package}
\alias{nexus-package}
\alias{nexus}
\title{nexus: Sourcing Archaeological Materials by Chemical Composition}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Exploration and analysis of compositional data in the framework of Aitchison (1986, ISBN: 978-94-010-8324-9). This package provides tools for chemical fingerprinting and source tracking of ancient materials.
}
\details{
\tabular{ll}{
\strong{Package:} \tab nexus \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.4.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{Zenodo:} \tab \doi{10.5281/zenodo.10225630} \cr
}
}
\section{Package options}{

\pkg{nexus} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{nexus.verbose}: a \code{\link{logical}} scalar. Should \R report extra information
on progress? Defaults to \code{\link[=interactive]{interactive()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/tesselle/nexus}
  \item \url{https://packages.tesselle.org/nexus/}
  \item Report bugs at \url{https://codeberg.org/tesselle/nexus/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
Anne Philippe \tab \emph{Université de Nantes, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
