% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OutlierIndex-class}
\alias{OutlierIndex-class}
\alias{.OutlierIndex}
\title{Outliers}
\description{
An S4 class to store the result of outlier detection.
}
\section{Slots}{

\describe{
\item{\code{samples}}{A \code{\link{character}} vector XXX.}

\item{\code{groups}}{A \code{\link{character}} vector XXX.}

\item{\code{distances}}{A \code{\link{numeric}} vector XXX.}

\item{\code{outliers}}{A \code{\link{logical}} vector giving the squared Mahalanobis distance.}

\item{\code{limit}}{An \code{\link{numeric}} vector XXX.}

\item{\code{robust}}{An \code{\link{logical}} vector XXX.}
}}

\seealso{
Other classes: 
\code{\link{CompositionMatrix-class}},
\code{\link{LogRatio-class}},
\code{\link{NumericMatrix-class}},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{classes}
