% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/outliers.R
\docType{methods}
\name{outliers}
\alias{outliers}
\alias{find_outliers}
\alias{find_outliers-method}
\alias{count_outliers}
\alias{count_outliers-method}
\alias{plot_outliers}
\alias{plot_outliers-method}
\alias{find_outliers,CompositionMatrix-method}
\alias{count_outliers,OutlierIndex-method}
\alias{plot_outliers,OutlierIndex,missing-method}
\alias{plot_outliers,OutlierIndex,CompositionMatrix-method}
\title{Outlier Detection}
\usage{
find_outliers(object, ...)

count_outliers(object)

plot_outliers(object, data, ...)

\S4method{find_outliers}{CompositionMatrix}(object, level = 0.975, robust = TRUE, alpha = 0.5)

\S4method{count_outliers}{OutlierIndex}(object)

\S4method{plot_outliers}{OutlierIndex,missing}(object)

\S4method{plot_outliers}{OutlierIndex,CompositionMatrix}(object, data, select = NULL)
}
\arguments{
\item{object}{An \linkS4class{OutlierIndex} object.}

\item{...}{Currently not used.}

\item{data}{A \linkS4class{CompositionMatrix} object.}

\item{level}{A length-one \code{\link{numeric}} vector giving the
significance level. \code{level} is used as a cut-off value for outlier
detection: observations with larger (squared) Mahalanobis distance are
considered as potential outliers.}

\item{robust}{A \code{\link{logical}} scalar: should robust estimators be
used (see \code{\link[robustbase:covMcd]{robustbase::covMcd()}})?}

\item{alpha}{A length-one \code{\link{numeric}} vector controlling the size of the
subsets over which the determinant is minimized (see
\code{\link[robustbase:covMcd]{robustbase::covMcd()}}). Only used if \code{robust} is \code{TRUE}.}

\item{select}{A \code{\link{numeric}} or \code{\link{character}} vector giving the selection of
the parts that are drawn.}
}
\value{
\code{find_outliers()} returns an \linkS4class{OutlierIndex} object.
}
\description{
Outlier Detection
}
\examples{
## Coerce to chemical data
data("kommos", package = "folio")
coda <- as_composition(kommos[, -c(1, 2)])
coda <- remove_NA(coda, margin = 1)

## Detect outliers
out <- find_outliers(coda)
count_outliers(out)

## Plot
plot_outliers(out) +
  ggplot2::theme_bw() +
  khroma::scale_colour_highcontrast()

plot_outliers(out, coda, select = 1:6) +
  khroma::scale_colour_YlOrBr(range = c(0.3, 1))
}
\references{
Filzmoser, P., Garrett, R. G. & Reimann, C. (2005). Multivariate outlier
detection in exploration geochemistry. \emph{Computers & Geosciences},
31(5), 579-587. \doi{10.1016/j.cageo.2004.11.013}.

Filzmoser, P. & Hron, K. (2008). Outlier Detection for Compositional Data
Using Robust Methods. \emph{Mathematical Geosciences}, 40(3), 233-248.
\doi{10.1007/s11004-007-9141-5}.

Filzmoser, P., Hron, K. & Reimann, C. (2012). Interpretation of multivariate
outliers for compositional data. \emph{Computers & Geosciences}, 39, 77-85.
\doi{10.1016/j.cageo.2011.06.014}.

Santos, F. (2020). Modern methods for old data: An overview of some robust
methods for outliers detection with applications in osteology. \emph{Journal of
Archaeological Science: Reports}, 32, 102423.
\doi{10.1016/j.jasrep.2020.102423}.
}
\seealso{
Other outlier detection methods: 
\code{\link{OutlierIndex-class}}
}
\author{
N. Frerebeau
}
\concept{outlier detection methods}
