% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plot_coda}
\alias{plot_coda}
\alias{autoplot,CompositionMatrix-method}
\alias{plot,CompositionMatrix,missing-method}
\title{Plot Compositional Data}
\usage{
\S4method{autoplot}{CompositionMatrix}(object, ..., order = NULL, decreasing = FALSE, facet = TRUE)

\S4method{plot}{CompositionMatrix,missing}(x, order = NULL, decreasing = FALSE, facet = TRUE, ...)
}
\arguments{
\item{...}{Currently not used.}

\item{order}{An \code{\link{integer}} vector giving the index of the column to be used
for the ordering of the data.}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort order be increasing
or decreasing?}

\item{facet}{A \code{\link{logical}} scalar: should a matrix of panels defined by
groups be drawn? Only used if \code{x} \link[=has_groups]{has groups defined}.}

\item{x, object}{A \linkS4class{CompositionMatrix} object.}
}
\value{
\itemize{
\item \code{autoplot()} returns a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
Displays a compositional bar chart.
}
\examples{
## Coerce to compositional data
data("hongite")
coda <- as_composition(hongite)

## Bar plot
plot(coda, order = 2)

## Use ceramic types for grouping
data("kommos", package = "folio")
kommos <- remove_NA(kommos, margin = 1) # Remove cases with missing values
coda <- as_composition(kommos, groups = 1)

plot(coda, order = 1)

## Log ratio
clr <- transform_clr(coda)
plot(clr, facet = FALSE)
}
\seealso{
Other plot methods: 
\code{\link{as_graph}()},
\code{\link{plot_logratio}}
}
\author{
N. Frerebeau
}
\concept{plot methods}
