% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plot_outliers}
\alias{plot_outliers}
\alias{plot,OutlierIndex,missing-method}
\title{Plot Outliers}
\usage{
\S4method{plot}{OutlierIndex,missing}(
  x,
  ...,
  type = c("dotchart", "qqplot"),
  pch.in = 16,
  pch.out = 3,
  ncol = NULL,
  flip = FALSE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  axes = TRUE,
  frame.plot = axes,
  panel.first = NULL,
  panel.last = NULL,
  legend = list(x = "top")
)
}
\arguments{
\item{x}{An \code{\linkS4class{OutlierIndex}} object.}

\item{...}{Further \link[graphics:par]{graphical parameters}.}

\item{type}{A \code{\link{character}} string specifying the type of plot that should be
made. It must be one of "\code{dotchart}" or "\code{qqplot}". Any unambiguous
substring can be given.}

\item{pch.in, pch.out}{A symbol specification for non-outliers and outliers
(resp.).}

\item{ncol}{An \code{\link{integer}} specifying the number of columns to use when
\code{facet} is "\code{multiple}". Defaults to 1 for up to 4 series, otherwise to 2.}

\item{flip}{A \code{\link{logical}} scalar: should the y-axis (ticks and numbering) be
flipped from side 2 (left) to 4 (right) from group to group?}

\item{xlab, ylab}{A \code{\link{character}} vector giving the x and y axis labels.}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}

\item{ann}{A \code{\link{logical}} scalar: should the default annotation (title and x
and y axis labels) appear on the plot?}

\item{axes}{A \code{\link{logical}} scalar: should axes be drawn on the plot?}

\item{frame.plot}{A \code{\link{logical}} scalar: should a box be drawn around the
plot?}

\item{panel.first}{An an \code{expression} to be evaluated after the plot axes are
set up but before any plotting takes place. This can be useful for drawing
background grids.}

\item{panel.last}{An \code{expression} to be evaluated after plotting has taken
place but before the axes, title and box are added.}

\item{legend}{A \code{\link{list}} of additional arguments to be passed to
\code{\link[graphics:legend]{graphics::legend()}}; names of the list are used as argument names.
If \code{NULL}, no legend is displayed.}
}
\value{
\code{plot()} is called for its side-effects: is results in a graphic being
displayed (invisibly return \code{x}).
}
\description{
Plot Outliers
}
\examples{
## Data from Day et al. 2011
data("kommos", package = "folio") # Coerce to compositional data
kommos <- remove_NA(kommos, margin = 1) # Remove cases with missing values
coda <- as_composition(kommos, groups = 1) # Use ceramic types for grouping

## Detect outliers
out <- outliers(coda, groups = NULL, robust = FALSE)

plot(out, type = "dotchart")
plot(out, type = "qqplot")

## Detect outliers by group
out <- outliers(coda[, 1:15])

plot(out, type = "dotchart", ncol = 2)
plot(out, type = "qqplot", ncol = 2)
}
\references{
Filzmoser, P., Garrett, R. G. & Reimann, C. (2005). Multivariate outlier
detection in exploration geochemistry. \emph{Computers & Geosciences},
31(5), 579-587. \doi{10.1016/j.cageo.2004.11.013}.

Filzmoser, P. & Hron, K. (2008). Outlier Detection for Compositional Data
Using Robust Methods. \emph{Mathematical Geosciences}, 40(3), 233-248.
\doi{10.1007/s11004-007-9141-5}.

Filzmoser, P., Hron, K. & Reimann, C. (2012). Interpretation of multivariate
outliers for compositional data. \emph{Computers & Geosciences}, 39, 77-85.
\doi{10.1016/j.cageo.2011.06.014}.
}
\seealso{
Other outlier detection methods: 
\code{\link{outliers}()}
}
\author{
N. Frerebeau
}
\concept{outlier detection methods}
