% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/outliers.R
\docType{methods}
\name{plot_outliers}
\alias{plot_outliers}
\alias{autoplot,OutlierIndex-method}
\alias{plot,OutlierIndex,missing-method}
\title{Plot Outliers}
\usage{
\S4method{autoplot}{OutlierIndex}(object, ..., qq = TRUE, limit = !qq)

\S4method{plot}{OutlierIndex,missing}(x, qq = TRUE, limit = !qq, ...)
}
\arguments{
\item{...}{Currently not used.}

\item{qq}{A \code{\link{logical}} scalar: should a quantile-quantile plot be produced?}

\item{limit}{A \code{\link{logical}} scalar: should the cut-off value for outlier
detection be displayed?}

\item{x, object}{An \linkS4class{OutlierIndex} object.}
}
\value{
\itemize{
\item \code{autoplot()} returns a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} and \code{qqplot()} are called it for their side-effects: they result
in a graphic being displayed (invisibly return \code{x}).
}
}
\description{
Plot Outliers
}
\examples{
## Coerce to chemical data
data("hongite")
coda <- as_composition(hongite)

## Detect outliers
out <- outliers(coda)

## Plot
plot(out, qq = TRUE)
plot(out, qq = FALSE)
}
\references{
Filzmoser, P., Garrett, R. G. & Reimann, C. (2005). Multivariate outlier
detection in exploration geochemistry. \emph{Computers & Geosciences},
31(5), 579-587. \doi{10.1016/j.cageo.2004.11.013}.

Filzmoser, P. & Hron, K. (2008). Outlier Detection for Compositional Data
Using Robust Methods. \emph{Mathematical Geosciences}, 40(3), 233-248.
\doi{10.1007/s11004-007-9141-5}.

Filzmoser, P., Hron, K. & Reimann, C. (2012). Interpretation of multivariate
outliers for compositional data. \emph{Computers & Geosciences}, 39, 77-85.
\doi{10.1016/j.cageo.2011.06.014}.
}
\seealso{
Other outlier detection methods: 
\code{\link{outliers}()}
}
\author{
N. Frerebeau
}
\concept{outlier detection methods}
