% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_composition}
\alias{as_features}
\alias{CompositionMatrix}
\alias{IncidenceMatrix}
\alias{replace_NA}
\alias{remove_NA}
\alias{remove_zero}
\alias{remove_empty}
\alias{get_groups}
\alias{has_groups}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{arkhe}{\code{\link[arkhe:CompositionMatrix-class]{CompositionMatrix}}, \code{\link[arkhe:IncidenceMatrix-class]{IncidenceMatrix}}, \code{\link[arkhe:coerce]{as_composition}}, \code{\link[arkhe:coerce]{as_features}}, \code{\link[arkhe:mutator]{get_groups}}, \code{\link[arkhe:mutator]{has_groups}}, \code{\link[arkhe:clean]{remove_NA}}, \code{\link[arkhe:clean]{remove_empty}}, \code{\link[arkhe:clean]{remove_zero}}, \code{\link[arkhe:clean]{replace_NA}}}
}}

