% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{replace_NA}
\alias{remove_NA}
\alias{remove_zero}
\alias{remove_empty}
\alias{biplot}
\alias{plot_contributions}
\alias{plot_cos2}
\alias{plot_individuals}
\alias{plot_variables}
\alias{plot_variance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{arkhe}{\code{\link[arkhe:missing]{remove_NA}}, \code{\link[arkhe:deprecate]{remove_empty}}, \code{\link[arkhe:zero]{remove_zero}}, \code{\link[arkhe:missing]{replace_NA}}}

  \item{dimensio}{\code{\link[dimensio]{biplot}}, \code{\link[dimensio]{plot_contributions}}, \code{\link[dimensio:plot_contributions]{plot_cos2}}, \code{\link[dimensio:plot_coordinates]{plot_individuals}}, \code{\link[dimensio:plot_coordinates]{plot_variables}}, \code{\link[dimensio:plot_eigenvalues]{plot_variance}}}
}}

