% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{samples}
\alias{samples}
\alias{has_replicates}
\alias{has_replicates-method}
\alias{get_samples}
\alias{get_samples-method}
\alias{set_samples<-}
\alias{set_samples-method}
\alias{has_replicates,CompositionMatrix-method}
\alias{has_replicates,LogRatio-method}
\alias{has_replicates,OutlierIndex-method}
\alias{get_samples,CompositionMatrix-method}
\alias{get_samples,LogRatio-method}
\alias{get_samples,OutlierIndex-method}
\alias{set_samples<-,CompositionMatrix-method}
\alias{set_samples,CompositionMatrix-method}
\title{Deal With Samples}
\usage{
has_replicates(x)

get_samples(x)

set_samples(x) <- value

\S4method{has_replicates}{CompositionMatrix}(x)

\S4method{has_replicates}{LogRatio}(x)

\S4method{has_replicates}{OutlierIndex}(x)

\S4method{get_samples}{CompositionMatrix}(x)

\S4method{get_samples}{LogRatio}(x)

\S4method{get_samples}{OutlierIndex}(x)

\S4method{set_samples}{CompositionMatrix}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set \code{samples}.}

\item{value}{A possible value for the \code{samples} of \code{x}.}
}
\value{
\itemize{
\item \code{set_samples()} returns an object of the same sort as \code{x} with the new
sample names assigned.
\item \code{get_samples()} returns the sample names of \code{x}.
\item \code{has_replicates()} returns a \code{\link{logical}} scalar.
}
}
\description{
Retrieves or defines the sample names.
}
\details{
In some situations, measurements may have been repeated (e.g. multiple
chemical analyses on the same sample). The presence of repeated
measurements can be specified by giving several observations the same
sample name.

See \code{vignette("nexus")}.
}
\seealso{
Other mutators: 
\code{\link{groups}},
\code{\link{subset}()},
\code{\link{totals}}
}
\author{
N. Frerebeau
}
\concept{mutators}
