% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slides}
\alias{slides}
\title{Thin Sections}
\format{
A \code{\link{data.frame}} with 9 variables:
\describe{
\item{analyst}{Analyst number.}
\item{slide}{Slide number.}
\item{quartz}{Quartz (percent).}
\item{microcline}{Microcline (percent).}
\item{plagioclass}{Plagioclass (percent).}
\item{biotite}{Biotite (percent).}
\item{plagioclass}{Plagioclass (percent).}
\item{muscovite}{Muscovite (percent).}
\item{opaques}{Opaque minerals (percent).}
\item{nonopaques}{Non-opaque minerals (percent).}
}
}
\usage{
slides
}
\description{
Mineral compositions of five slides as reported by five analysts.
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall. \doi{10.1007/978-94-009-4109-0}.
}
\seealso{
Other datasets: 
\code{\link{arctic}},
\code{\link{chemistry}},
\code{\link{hongite}},
\code{\link{petrography}}
}
\concept{datasets}
\keyword{datasets}
