% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/subset.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{[,CompositionMatrix-method}
\alias{[<-,CompositionMatrix-method}
\alias{[[<-,CompositionMatrix-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[}{CompositionMatrix}(x, i, j, ..., drop = TRUE)

\S4method{[}{CompositionMatrix}(x, i, j, ...) <- value

\S4method{[[}{CompositionMatrix}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An object from which to extract element(s) or in which to replace
element(s) (typically a \verb{*Matrix} object).}

\item{i, j}{Indices specifying elements to extract or replace. Indices are
\code{\link{numeric}}, \code{\link{integer}} or \code{\link{character}} vectors or empty (missing) or
\code{NULL}. Numeric values are coerced to \code{\link{integer}} as by \code{\link[=as.integer]{as.integer()}}
(and hence truncated towards zero). Character vectors will be matched to
the name of the elements. An empty index (a comma separated blank) indicates
that all entries in that dimension are selected.}

\item{...}{Currently not used.}

\item{drop}{A \code{\link{logical}} scalar: should the result be coerced to
the lowest possible dimension? This only works for extracting elements,
not for the replacement.}

\item{value}{A possible value for the element(s) of \code{x}.}
}
\value{
A subsetted object of the same sort as \code{x}.
}
\description{
Operators acting on objects to extract or replace parts.
}
\examples{
## Coerce to compositonal data
data("hongite")
coda <- as_composition(hongite)

## codaccess
dim(coda) # Get the matrix dimensions
row(coda) # Get the row indexes
col(coda, as.factor = TRUE) # Get the column indexes
nrow(coda) # Get the number of rows
ncol(coda) # Get the number of columns
dimnames(coda) # Get the dimension names
rownames(coda) <- LETTERS[1:25] # Set the row names
rownames(coda) # Get the rownames
colnames(coda) <- letters[21:25] # Set the column names
colnames(coda) # Get the column names

## Subset
coda[[1, 1]] # Get the first value
coda[1] # Get the first value
coda[, ] # Get all values
coda[1, , drop = FALSE] # Get the first row
coda[, 1:3] # Get the first three column
}
\seealso{
Other mutators: 
\code{\link{groups}},
\code{\link{samples}}
}
\author{
N. Frerebeau
}
\concept{mutators}
