% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/transform.R
\docType{methods}
\name{transform_alr}
\alias{transform_alr}
\alias{transform_alr-method}
\alias{transform_alr,CompositionMatrix-method}
\title{Additive Log-Ratios (ALR)}
\usage{
transform_alr(object, ...)

\S4method{transform_alr}{CompositionMatrix}(object, j = ncol(object))
}
\arguments{
\item{object}{A \linkS4class{CompositionMatrix} object.}

\item{...}{Currently not used.}

\item{j}{An \code{\link{integer}} giving the index of the rationing part (denominator).}
}
\value{
An \linkS4class{ALR} object.
}
\description{
Computes ALR transformation.
}
\examples{
## Coerce to chemical data
data("hongite")
coda <- as_composition(hongite)

## Centered log-ratio
clr <- transform_clr(coda)

## Additive log-ratio
alr <- transform_alr(coda)

## Isometric log-ratio
ilr1 <- transform_ilr(coda)
ilr2 <- transform_plr(coda)
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall. \doi{10.1007/978-94-009-4109-0}.

Greenacre, M. J. (2019). \emph{Compositional Data Analysis in Practice}.
Boca Raton: CRC Press.
}
\seealso{
Other log-ratio transformations: 
\code{\link{LogRatio}},
\code{\link{transform_clr}()},
\code{\link{transform_ilr}()},
\code{\link{transform_inverse}()},
\code{\link{transform_plr}()}
}
\author{
N. Frerebeau
}
\concept{log-ratio transformations}
