% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{variation}
\alias{variation}
\alias{variation-method}
\alias{variation,CompositionMatrix-method}
\title{Variation Matrix}
\usage{
variation(object, ...)

\S4method{variation}{CompositionMatrix}(object)
}
\arguments{
\item{object}{A \linkS4class{CompositionMatrix} object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{matrix}}.
}
\description{
Computes the compositional variation array.
}
\details{
The compositional variation array is a square matrix where the upper
triangular part displays the log-ratio variances and the lower triangular
part displays the log-ratio means.
}
\examples{
## Coerce to compositonal data
data("hongite")
coda <- as_composition(hongite)

## Variance matrix
var(coda)

## Covariance matrix
cov(coda)

## Variation matrix
variation(coda)
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall, p. 64-91. \doi{10.1007/978-94-009-4109-0}.
}
\seealso{
Other statistics: 
\code{\link{covariance}},
\code{\link{distance}}
}
\author{
N. Frerebeau
}
\concept{statistics}
